完善和修复渲染器，修复STB错误
渲染器：
添加 SamplerState 类
添加 Graphics 命名空间
将所有的 图形枚举迁移到 Graphics 命名空间
添加 Helper 命名空间提供各种工具方法

STB：
将构建配置定义的 STB_IMAGE_IMPLEMENTATION 宏转移到main.cpp

+++ core/scene/panel/hex_manager.h
@@ -39,10 +39,10 @@ namespace HexMetrics
 class HexManager
 {
 public:
-    float hexWidth, hexHeight;
+    int hexWidth, hexHeight;
     int tileWidth, tileHeight, sideLength;
     float outerRadius, innerRadius;
-    HexManager(float hexWidth, float hexHeight, int tileWidth, int tileHeight);
+    HexManager(int hexWidth, int hexHeight, int tileWidth, int tileHeight);
 
     [[nodiscard]] glm::vec2 HexToPixel(const glm::ivec2& hex) const;
     [[nodiscard]] glm::ivec2 PixelToHex(const glm::vec2& pixel) const;
